basic_programming/lesson_31/code/dz20240524_1/test/MainTest.java
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class MainTest {


    @Test
    @DisplayName("parsePerson: regular case")
    void parsePersons() {
        List<String> data = List.of("John,18",
                "Jack,32",
                "Anna,21",
                "Olga,25",
                "Gektar,21"
        );
        List<Person> expectedResult = List.of(
                new Person("John",18),
                new Person("Jack",32),
                new Person("Anna",21),
                new Person("Olga",25),
                new Person("Gektar",21)
        );

        List<Person> actualResult = Main.parsePersons(data);

        Assertions.assertEquals(expectedResult, actualResult);
    }

    @Test
    @DisplayName("parsePerson: input empty list")
    void testParsePersons() {
        //List<String> data = Collections.EMPTY_LIST;
        List<String> data = new ArrayList<>();
        List<Person> expectedResult = new ArrayList<>();

        List<Person> actualResult = Main.parsePersons(data);

        Assertions.assertEquals(expectedResult, actualResult);

    }

    @Test
    @DisplayName("parsePerson: input null")
    void parsePersons_List_Null() {
        //List<String> data = Collections.EMPTY_LIST;
        List<String> data = null;
        List<Person> expectedResult = new ArrayList<>();

        List<Person> actualResult = Main.parsePersons(data);

        Assertions.assertEquals(expectedResult, actualResult);
    }

    @Test
    @DisplayName("parsePerson: list with null and empty values")
    void parsePersons_If_listWithNullAndEmptyValues() {
        List<String> data = new ArrayList<>(List.of(
                "John,18",
                "",
                "Jack,32",
                "Anna,21",
                "Olga,25",
                "Jack32",
                "Gektar,21"
        ));
        data.add(null);
        List<Person> expectedResult = List.of(
                new Person("John",18),
                new Person("Jack",32),
                new Person("Anna",21),
                new Person("Olga",25),
                new Person("Gektar",21)
        );

        List<Person> actualResult = Main.parsePersons(data);

        Assertions.assertEquals(expectedResult, actualResult);
    }
}